refator: move site-related tools to the site directory independently.
