+++ src/tools/site/record.ts
@@ -1,37 +1,10 @@
 import { CallToolRequest, Tool } from '@modelcontextprotocol/sdk/types.js';
-import api from '../utils/service.js';
+import api from '../../utils/service.js';
 import {
-  CreateRecordRequest,
   ListRecordsRequest,
-  ListSitesRequest,
-  CreateSiteRequest,
+  CreateRecordRequest,
 } from '@alicloud/esa20240910';
-import { GetMatchSiteRequest } from '../utils/types.js';
 
-export const SITE_ACTIVE_LIST_TOOL: Tool = {
-  name: 'site_active_list',
-  description: 'List all active sites',
-  inputSchema: {
-    type: 'object',
-    properties: {},
-  },
-};
-
-export const SITE_MATCH_TOOL: Tool = {
-  name: 'site_match',
-  description:
-    'Identify which site in the account matches the provided input criteria.',
-  inputSchema: {
-    type: 'object',
-    properties: {
-      recordName: {
-        type: 'string',
-        description: 'The name of the site to match',
-      },
-    },
-    required: ['recordName'],
-  },
-};
 
 export const SITE_RECORD_LIST_TOOL: Tool = {
   name: 'site_record_list',
@@ -48,8 +21,8 @@ export const SITE_RECORD_LIST_TOOL: Tool = {
   },
 };
 
-export const CREATE_A_OR_AAAA_RECORD_TOOL: Tool = {
-  name: 'create_a_or_aaaa_record',
+export const CREATE_SITE_A_OR_AAAA_RECORD_TOOL: Tool = {
+  name: 'create_site_a_or_aaaa_record',
   description:
     'Creates a DNS record for a specific website. Only A/AAAA records are supported.',
   inputSchema: {
@@ -116,8 +89,8 @@ export const CREATE_A_OR_AAAA_RECORD_TOOL: Tool = {
   },
 };
 
-export const CREATE_CNAME_RECORD_TOOL: Tool = {
-  name: 'create_cname_record',
+export const CREATE_SITE_CNAME_RECORD_TOOL: Tool = {
+  name: 'create_site_cname_record',
   description:
     'Creates a DNS record for a specific website. Only supports records with type=CNAME and sourceType=Domain.',
   inputSchema: {
@@ -188,8 +161,8 @@ export const CREATE_CNAME_RECORD_TOOL: Tool = {
   },
 };
 
-export const CREATE_TXT_RECORD_TOOL: Tool = {
-  name: 'create_txt_record',
+export const CREATE_SITE_TXT_RECORD_TOOL: Tool = {
+  name: 'create_site_txt_record',
   description:
     'Creates a DNS record for a specific website. Only TXT records are supported.',
   inputSchema: {
@@ -239,56 +212,8 @@ export const CREATE_TXT_RECORD_TOOL: Tool = {
   },
 };
 
-export const CREATE_SITE_TOOL: Tool = {
-  name: 'create_site',
-  description:
-    'Adds a website. Make sure that you have an available plan before you add a website. Make sure that your website domain name has an ICP filing if the location you want to specify covers the Chinese mainland.',
-  inputSchema: {
-    type: 'object',
-    properties: {
-      siteName: {
-        type: 'string',
-        description: 'The website name.',
-        examples: ['example.com'],
-      },
-      coverage: {
-        type: 'string',
-        description:
-          'The service location. Valid values:\n- domestic: the Chinese mainland\n- global: global\n- overseas: outside the Chinese mainland',
-        enum: ['global', 'domestic', 'overseas'],
-        examples: ['domestic'],
-      },
-      accessType: {
-        type: 'string',
-        description: 'The DNS setup. Valid values:\n- NS;\n- CNAME',
-        enum: ['NS', 'CNAME'],
-        examples: ['NS'],
-      },
-      instanceId: {
-        type: 'string',
-        description:
-          'The instance ID, which can be obtained by calling the [ListUserRatePlanInstances] operation. Specify at least one of the instance ID and website ID. If you specify both of them, the instance ID is used.',
-        examples: ['dbaudit-cn-nwy349jdb03'],
-      },
-      resourceGroupId: {
-        type: 'string',
-        description:
-          'The ID of the resource group. If you leave this parameter empty, the system uses the default resource group ID.',
-        examples: ['rg-acfmw4znnok****'],
-      },
-    },
-    required: ['siteName', 'coverage', 'accessType', 'instanceId'],
-    annotations: {
-      readOnlyHint: false,
-      destructiveHint: false,
-      idempotentHint: false,
-      openWorldHint: false,
-    },
-  },
-};
-
-export const CREATE_NS_RECORD_TOOL: Tool = {
-  name: 'create_ns_record',
+export const CREATE_SITE_NS_RECORD_TOOL: Tool = {
+  name: 'create_site_ns_record',
   description:
     'Creates a DNS record for a specific website. Only NS records are supported.',
   inputSchema: {
@@ -298,47 +223,48 @@ export const CREATE_NS_RECORD_TOOL: Tool = {
         type: 'number',
         description:
           'The website ID, which can be obtained by calling the [ListSites] operation.',
-      recordName: {
-        type: 'string',
-        description: 'The record name.',
-        examples: ['www.example.com'],
-      },
-      ttl: {
-        type: 'number',
-        description:
-          'The TTL of the record. Unit: seconds. If the value is 1, the TTL of the record is determined by the system. Default: 1.',
-        examples: [1],
-      },
-      data: {
-        type: 'object',
-        description:
-          'The DNS record information. The format of this field varies based on the record type. For more information, see https://www.alibabacloud.com/help/doc-detail/2708761.html',
-        properties: {
-          value: {
-            type: 'string',
-            description: 'The name servers for the domain name. Required.',
-            example: ['ns.example.com'],
+        recordName: {
+          type: 'string',
+          description: 'The record name.',
+          examples: ['www.example.com'],
+        },
+        ttl: {
+          type: 'number',
+          description:
+            'The TTL of the record. Unit: seconds. If the value is 1, the TTL of the record is determined by the system. Default: 1.',
+          examples: [1],
+        },
+        data: {
+          type: 'object',
+          description:
+            'The DNS record information. The format of this field varies based on the record type. For more information, see https://www.alibabacloud.com/help/doc-detail/2708761.html',
+          properties: {
+            value: {
+              type: 'string',
+              description: 'The name servers for the domain name. Required.',
+              example: ['ns.example.com'],
+            },
           },
         },
+        comment: {
+          type: 'string',
+          description:
+            'The comment of the record. The maximum length is 100 characters.',
+        },
       },
-      comment: {
-        type: 'string',
-        description:
-          'The comment of the record. The maximum length is 100 characters.',
+      required: ['siteId', 'recordName', 'type', 'data'],
+      annotations: {
+        readOnlyHint: false,
+        destructiveHint: false,
+        idempotentHint: false,
+        openWorldHint: false,
       },
-    },
-    required: ['siteId', 'recordName', 'type', 'data'],
-    annotations: {
-      readOnlyHint: false,
-      destructiveHint: false,
-      idempotentHint: false,
-      openWorldHint: false,
     },
   },
 };
 
-export const CREATE_MX_RECORD_TOOL: Tool = {
-  name: 'create_mx_record',
+export const CREATE_SITE_MX_RECORD_TOOL: Tool = {
+  name: 'create_site_mx_record',
   description:
     'Creates a DNS record for a specific website. Only MX records are supported.',
   inputSchema: {
@@ -414,41 +340,9 @@ export const site_record_list = async (request: CallToolRequest) => {
   };
 };
 
-export const site_active_list = async () => {
-  const res = await api.listSites({
-    siteSearchType: 'fuzzy',
-    status: 'active',
-    pageNumber: 1,
-    pageSize: 500,
-  } as ListSitesRequest);
-  return {
-    content: [
-      {
-        type: 'text',
-        text: JSON.stringify(res),
-      },
-    ],
-    success: true,
-  };
-};
-
-export const site_match = async (request: CallToolRequest) => {
-  const res = await api.getMatchSite({
-    recordName: request.params.arguments?.recordName ?? '',
-  } as GetMatchSiteRequest);
-  return {
-    content: [
-      {
-        type: 'text',
-        text: JSON.stringify(res),
-      },
-    ],
-    success: true,
-  };
-};
-
-
-export const create_a_or_aaaa_record = async (request: CallToolRequest) => {
+export const create_site_a_or_aaaa_record = async (
+  request: CallToolRequest,
+) => {
   const req = request.params.arguments as CreateRecordRequest;
 
   req.proxied = req.proxied || false;
@@ -470,7 +364,7 @@ export const create_a_or_aaaa_record = async (request: CallToolRequest) => {
   };
 };
 
-export const create_cname_record = async (request: CallToolRequest) => {
+export const create_site_cname_record = async (request: CallToolRequest) => {
   const req = request.params.arguments as CreateRecordRequest;
 
   req.type = req.type || 'CNAME';
@@ -499,7 +393,7 @@ export const create_cname_record = async (request: CallToolRequest) => {
   };
 };
 
-export const create_txt_record = async (request: CallToolRequest) => {
+export const create_site_txt_record = async (request: CallToolRequest) => {
   const req = request.params.arguments as CreateRecordRequest;
 
   req.type = req.type || 'TXT';
@@ -520,7 +414,7 @@ export const create_txt_record = async (request: CallToolRequest) => {
   };
 };
 
-export const create_ns_record = async (request: CallToolRequest) => {
+export const create_site_ns_record = async (request: CallToolRequest) => {
   const req = request.params.arguments as CreateRecordRequest;
 
   req.type = req.type || 'NS';
@@ -541,7 +435,7 @@ export const create_ns_record = async (request: CallToolRequest) => {
   };
 };
 
-export const create_mx_record = async (request: CallToolRequest) => {
+export const create_site_mx_record = async (request: CallToolRequest) => {
   const req = request.params.arguments as CreateRecordRequest;
 
   req.type = req.type || 'MX';
@@ -555,17 +449,6 @@ export const create_mx_record = async (request: CallToolRequest) => {
   }
 
   const res = await api.createRecord(req);
-
-  return {
-    content: [{ type: 'text', text: JSON.stringify(res) }],
-    success: true,
-  };
-};
-
-export const create_site = async (request: CallToolRequest) => {
-  const res = await api.createSite(
-    request.params.arguments as CreateSiteRequest,
-  );
 
   return {
     content: [{ type: 'text', text: JSON.stringify(res) }],
